samples/microsoft/python/getting-started-agents/hosted-agents/mcp-agent/requirements.txt
azure-ai-agentserver-agentframework==1.0.0b4
agent-framework

pytest==8.4.2
azure-identity==1.25.0
python-dotenv==1.1.1
azure-monitor-opentelemetry==1.8.1